Working stage 4 fixed empty nodes leading up to prefix not deleting when deleteAllWithPrefix is called. Fixed issue with document not deleting from documentStore when Trie deletes are called

+++ DataStructures/project/Stage4/src/main/java/edu/yu/cs/com1320/project/impl/TrieImpl.java
@@ -5,13 +5,13 @@ import edu.yu.cs.com1320.project.Trie;
 import java.util.*;
 
 public class TrieImpl<Value> implements Trie<Value> {
-    private final int alphabetSize = 256; // extended ASCII
+    private static final int alphabetSize = 256; // extended ASCII
     private final Node<Value> root; // root of trie
     private int wordsInTrie;
 
-    private class Node<Value> {
+    private static class Node<Value> {
         private String valueString;
-        protected Set<Value> documents = new HashSet<>();
+        protected Set<Value> values = new HashSet<>();
         protected Node[] links = new Node[alphabetSize];
         private boolean shouldDelete = false;
         private boolean isWord = false;
@@ -49,7 +49,7 @@ public class TrieImpl<Value> implements Trie<Value> {
         if(d == key.length()){
             node.valueString = key;
             node.isWord = true;
-            node.documents.add(value);
+            node.values.add(value);
             return node;
         }
         char c = key.charAt(d);
@@ -67,8 +67,9 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public Set<Value> get(String key) {
+        if(key == null) throw new IllegalArgumentException("key is null");
         Node x = this.get(this.root, key, 0);
-        return (x == null) ? new HashSet<>() : x.documents;
+        return (x == null) ? new HashSet<>() : x.values;
     }
 
     private Node get(Node x, String key, int d) {
@@ -98,6 +99,7 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public List<Value> getSorted(String key, Comparator<Value> comparator) {
+        if(key == null) throw new IllegalArgumentException("key is null");
         List<Value> listToReturn = new ArrayList<>(get(key).stream().toList());
         listToReturn.sort(comparator);
         return listToReturn;
@@ -115,12 +117,14 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public List<Value> getAllWithPrefixSorted(String prefix, Comparator<Value> comparator) {
+        if(prefix == null) throw new IllegalArgumentException("prefix is null");
         Node prefixNode = getNode(this.root, prefix, 0);
-        assert prefixNode != null;
+        //assert prefixNode != null;
         Set<Value> documentSet = new HashSet<>();
         documentSet = this.getPrefixDocumentSet(prefixNode, documentSet);
         List<Value> documentListToReturn = new ArrayList<>(documentSet);
         documentListToReturn.sort(comparator);
+
         return documentListToReturn;
     }
     private Node getNode(Node node, String key, int depth){
@@ -132,7 +136,7 @@ public class TrieImpl<Value> implements Trie<Value> {
 
     private Set<Value> getPrefixDocumentSet(Node node, Set<Value> documentSet){
         if(node == null) return documentSet;
-        if(!node.documents.isEmpty()) documentSet.addAll(node.documents);
+        if(!node.values.isEmpty()) documentSet.addAll(node.values);
         if(Arrays.stream(node.links).allMatch(Objects::isNull)) return documentSet;
         for (Node childNode : node.links){
             getPrefixDocumentSet(childNode, documentSet);
@@ -148,18 +152,20 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public Set<Value> deleteAllWithPrefix(String prefix) {
+        if(prefix == null) throw new IllegalArgumentException("prefix is null");
         Set<Value> deletedDocs = new HashSet<>();
         Node subTrie = getNode(this.root, prefix, 0);
         if(subTrie != null) {
             deleteRootWithPrefix(subTrie, deletedDocs);
         }
-        //deleteAllEmptyLeafs(this.root);
+        deleteAll(prefix); //deletes all empty nodes up to prefix
         return deletedDocs;
     }
 
     private Set<Value> deleteRootWithPrefix(Node node, Set<Value> deletedDocs){
         if (node == null) return deletedDocs;
-        deletedDocs.addAll(node.documents);
+        //fix the issue with empty nodes not deleting
+        deletedDocs.addAll(node.values);
         for (Node childNode : node.links){
             if(childNode != null) {
                 deleteRootWithPrefix(childNode, deletedDocs);
@@ -178,6 +184,7 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public Set<Value> deleteAll(String key) {
+        if(key == null) throw new IllegalArgumentException("key is null");
         Set<Value> setToReturn = new HashSet<>();
         deleteAll(this.root, key, 0, setToReturn);
         //deleteAllEmptyLeafs(this.root);
@@ -187,15 +194,15 @@ public class TrieImpl<Value> implements Trie<Value> {
     private Node deleteAll(Node node, String endKey, int depth, Set<Value> setToReturn) {
        if(node == null) return null;
        if(depth == endKey.length()){
-           setToReturn.addAll(node.documents);
-           node.documents.clear();
-           if(!node.hasLinks()) node.documents = null;
+           setToReturn.addAll(node.values);
+           node.values.clear();
+           if(!node.hasLinks()) node.values = null;
            wordsInTrie--;
        } else {
            char currentChar = endKey.charAt(depth);
            node.links[currentChar] = this.deleteAll(node.links[currentChar], endKey, depth+1, setToReturn);
        }
-        if(node.documents != null && !node.documents.isEmpty()) return node;
+        if(node.values != null && !node.values.isEmpty()) return node;
         for (Node childNode : node.links) {
             if (childNode != null) {
                 return node;
@@ -204,22 +211,7 @@ public class TrieImpl<Value> implements Trie<Value> {
         return null;
     }
 
-    private void deleteAllEmptyLeafs(Node node){
-        if(!node.hasLinks() && node.documents.isEmpty()) {
-            node.shouldDelete = true;
-            return;
-        }
-        if(node.hasLinks()){
-            for (Node childNode : node.links){
-                if (childNode != null && childNode.hasLinks()){
-                    deleteAllEmptyLeafs(childNode);
-                    if(childNode.shouldDelete) {
-                        childNode = null;
-                    }
-                }
-            }
-        }
-    }
+
 
 
 
@@ -232,6 +224,7 @@ public class TrieImpl<Value> implements Trie<Value> {
      */
     @Override
     public Value delete(String key, Value val) {
+        if(key == null || val == null) throw new IllegalArgumentException("key or value is null");
         return deleteValue(this.root, key, val, 0);
     }
 
@@ -240,9 +233,9 @@ public class TrieImpl<Value> implements Trie<Value> {
             return null;
         }
         if (depth == key.length()) {
-            if (node.documents.remove(val)) {
+            if (node.values.remove(val)) {
                 // if there are no more documents in the documentSet of this node decrease wordsInTrie
-                if (node.documents.isEmpty()) {
+                if (node.values.isEmpty()) {
                     wordsInTrie--;
                     //return val;
                 }
@@ -257,7 +250,7 @@ public class TrieImpl<Value> implements Trie<Value> {
         Value valueToReturn = (Value) deleteValue(node.links[currentChar], key, val, depth + 1);
 
         //extract the deleted value and if no links or docs set link to null
-        if (valueToReturn != null && node.links[currentChar].documents.isEmpty() && !node.links[currentChar].hasLinks()) {
+        if (valueToReturn != null && node.links[currentChar].values.isEmpty() && !node.links[currentChar].hasLinks()) {
             node.links[currentChar] = null;
         }
